move ModelServiceTest to services; make a constant
