+++ src/test/java/mops/rheinjug2/services/ModelServiceTest.java
@@ -1,4 +1,4 @@
-package mops.rheinjug2.service;
+package mops.rheinjug2.services;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
@@ -11,7 +11,6 @@ import mops.rheinjug2.entities.Event;
 import mops.rheinjug2.entities.Student;
 import mops.rheinjug2.repositories.EventRepository;
 import mops.rheinjug2.repositories.StudentRepository;
-import mops.rheinjug2.services.ModelService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
